refactor worker writing task dispatch
